Return float as float

+++ test_templates/inputs_n_outputs/template/take_input_produce_output.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
         'bool_result': True,
         'false_result': False,
         'number_1_result': int(os.environ['NUMBER_1']),
-        'number_2_result': int(os.environ['NUMBER_2'])}
+        'number_2_result': float(os.environ['NUMBER_2'])}
     }
 
     with open(os.environ['RESULT_FILE'], 'w', encoding="utf8") as target_file:
